feat: Cleaned up code. Extracted out muzero from main along with some helper functions. asserted all functions in main work. ray dynamics function pretraining works. fixed some typos. Expanded readme
