+++ config.py
@@ -11,23 +11,22 @@ class Config:
     SGD_epoch = 4
     update_every = 4
     learning_update = 0
-    action_space = 100
     # Mcts
-    nodes_per_turn = {
-        0: 100,
-        1: 500,
-        2: action_space,
-        3: action_space,
-        4: action_space,
-        5: action_space,
-    }
+    # nodes_per_turn = {
+    #     0: 100,
+    #     1: 500,
+    #     2: action_space,
+    #     3: action_space,
+    #     4: action_space,
+    #     5: action_space,
+    # }
     node_policy_random_split = 0.75
     # Optimizer
     beta_1 = 0.9
     beta_2 = 0.95
     L2 = 0.1
     # WORDLE
-    word_restriction = None
+    word_restriction = 10
     # PER BUFFER
     batch_size = 4096
     num_unroll_steps = 6
@@ -42,7 +41,7 @@ class Config:
     checkpoint_interval = 5
     discount_rate = 0.85
     epsilon = 0.5
-    action_space = 2309
+    action_space = 10  # 2309
     value_loss_weight = 0.25
     weight_decay = 0.1
     lr_init = 3e-3
@@ -56,7 +55,7 @@ class Config:
     max_simulations = 100
     revisit_policy_search_rate = 0
     self_play_delay = 0  # Number of seconds to wait after each played game
-    training_delay = 5  # Number of seconds to wait after each training step
+    training_delay = 0  # Number of seconds to wait after each training step
     ratio = None  # Desired training steps per self played step ratio. Equivalent to a synchronous version, training can take much longer. Set it to None to disable it
     temperature_threshold = None  # Number of moves before dropping the temperature given by visit_softmax_temperature_fn to 0 (ie selecting the best action). If None, visit_softmax_temperature_fn is used every time
     num_workers = 8  # Number of simultaneous threads/workers self-playing to feed the replay buffer
@@ -69,7 +68,7 @@ class Config:
     load_dynamic_weights = False
     # paths
     buffer_path = (
-        pathlib.Path(__file__).resolve().parents[0] / "dataset"
+        pathlib.Path(__file__).resolve().parents[0] / "replay_buffer_storage"
     )  # Path to TensorBoard logs
     weights_path = (
         pathlib.Path(__file__).resolve().parents[0]
